Extract formatDateTime to separate module, refactor
